[cm] Allow dynamic children as sources/targets.
In cm_rust, `OfferSource` and `OfferTargets` s can only be static
children, not dynamic children in collections. This means most of the
component manager code base can't even model a dynamic offer. This
change makes it so `OfferSources` and `OfferTargets` can have a
Collection.

I don't love this solution. It introduces many places in the code
where an offer is known to have a static child (i.e., any offers that
are defined in ComponentDecls), but the type system now forgets that
fact. It'd be nice if there was a separate set of types for
statically-declared offers as opposed to offers as they exist at
runtime. Unfortunately, `OfferDecl` is used to mean both a static
offer and a runtime offer throughout the CF codebase, and changing
that would be a much larger project.

This CL aims to be as "yadda yadda" as possible. Dynamic children are
still not supported, and if a ChildRef is found with `collection` set,
the code will panic. Actually addressing the dynamic cases will come
in further CLs.

Bug: 81207

Change-Id: Ic26f4bfcb925d2c90eaf4f4a40cb3c486e87253d

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544381
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>