[settings] Use executor for camera watcher test
Changes the test_camera_devices_watcher_timeout test to use an executor
with fake time so that it does not have to wait for the real timeout to
complete the test.

Also factors out the executor evaluation code to a new test helpers file
for reuse in other tests.

Fixed: 82500
Test: fx test setui_service_tests

Change-Id: I375d1de5c4ab45471d2984d582383891b1b6049d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579052
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Sijie Chen <sijiec@google.com>
Reviewed-by: William Xiao <wxyz@google.com>