[fx][scripts][fish] Fix completions when there is no vendor/ directory
When cloning the repository, there is no vendor/ directory. As per
https://fishshell.com/docs/current/language.html?highlight=exceptions#wildcards-globbing,
this leads to an error when expanding the glob $FUCHSIA_DIR/vendor/*/scripts/devshell/.

This change uses `set` for the glob, which does not lead to errors if
the result is empty.

Change-Id: I3a1f61188297835ba7acf1f9aa9bc2b502a87958
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578942
Reviewed-by: Anthony DiGirolamo <tonymd@google.com>
Commit-Queue: Grégoire Geis <gregoiregeis@google.com>