[testrunner] Make parseOutKernelReader simpler.
There are some flakes where the output from this reader doesn't match
what shows up in the serial log. I could not reproduce the flakes, but
I made some changes to simplify the reader to make it easier to follow.

I made the following changes:
- set new variables within the goroutine calling Read() and pass them as
a struct to the channel that we're waiting on. Previously, the goroutine
was setting variables that were also getting set outside the goroutine
which could possibly cause a race issue.
- work in []byte instead of switching back and forth from []byte to
string to []byte
- call the underlying readers's Read() at most once in a single call to
parseOutKernelReader.Read(). Before, we would keep reading until we
filled the length of buf or there were no more bytes to read.

Bug: 83608
Change-Id: I9db72eb21f432414b448caa7ad0fa654ec73b215
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579742
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>