[starnix] Update socket to read from pipe
Previously the socket would write to the pipe that it owned, but this
change flips that around so that a socket reads from the pipe that it
owns and writes to the pipe of the connected socket.

This makes the read/write file logic a bit more sensible, since we can
read out remaining bytes even after the connected socket has been
disconnected.

Change-Id: I994095a0fc13f45e724170e839ef0927de9d15a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579961
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>