[wlan][iwlwifi] remove iwl_mvm_update_quotas_single_iface()
The TIME_QUOTA_CMD crashed the firmware while disassociation
on 8265 and 9260. However, this command is not mandatory.
Since this function only sets the default value in the single
interface situation (the only mode we currently support), we
can safely remove it. A TODO(43218) is already filed to support
multiple interfaces.

Bug: 81770
Tested: on 8265 and 9260, 100 * (assoc + disassoc)

Change-Id: I87326b20ec73d9978d1d21a371f77482c47c2ac2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573481
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>