[audio][tests] Specify Impulse channels
With this CL, the HermeticImpulseTest::TestCase can optionally
specify which channels should be validated for the presence of
the expected impulse. Previously, we would expect this from every
output channel, and this remains the default behavior, if the
"channels_to_test" field is unspecified.

This is part one of a two-repo change, to re-enable pipeline
Impulse testing on Sherlock.

Change-Id: I139ef200f51213c9802da8660acb7ac33d1f18ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579906
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>