[fidlc] Split up new_syntax_tests.cc
During the syntax migration work, it was convenient to have a file
to dump tests that applied specifically to the new syntax. It no
longer makes sense to have this file, since it makes it harder to
know where a feature is tested. This CL splits up the test cases
from this file into the appropriate feature-based test files,
ensuring that no test coverage is lost in the process.

A large number of tests in this file exercised the layout parameter
and type constraint code which is new to the new syntax; these
tests have all been moved to types_tests.cc since they relate
to type compilation

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Change-Id: I57b83792a099a7024122f482787efb02bfac1240
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577561
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Fuchsia-Auto-Submit: Felix Zhu <fcz@google.com>