[fidlc] Remove derived names
Derived names represent names that correspond to part of a source
file (i.e. have a non-virtual SourceSpan), but that have a separately
specifed name.

This can be considered a more general form of anonymous names, where
the source span refers to the definition of the anonymous object,
and the manually specified name is the flattened name that is reserved
by fidlc.

Since there are no derived names that are not anonymous names, the
code for derived names is removed; things are not renamed to from
"anonymous" to "derived" since "anonymous" is considered to be
more explicit about the use case.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler --gtest_filter='TypesTests.*'

Change-Id: I03dcca2489736b592b1b7ee094fbd2fff48cbef9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546662
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Felix Zhu <fcz@google.com>