[clock][testrunner] Enforce outer test timeout.
Set a slightly longer outer timeout for running a single test so that
the tester can handle the test timeout in its own way, but if it fails
to, then we can fall back to the outer timeout.

We used to use a context timeout to do this, but it was removed in
commit 2315fc48e07c6b861190e6029148410bf9999b5e because it was hitting
this timeout just before it could hit the tester timeout and let the
tester do any necessary cleanup. However, there is still a possibility
that the target's timeout mechanism (which we currently rely on) can
fail, so we should have a larger backup timeout as well.

Bug: 84445
Change-Id: Ib72da1700bdcac4d6a4e944357272bcd277bc549
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580281
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Boone <gboone@google.com>