[overnet] Map Error::Done to Ok(0)
Due to this change in quiche
https://github.com/cloudflare/quiche/commit/95ea1361794e533bdb75599246f2f92033720621
the future implementation of the quic protocol will return
Err(Error::Done) in some cases where it used to return Ok(0). Overnet
relies on the Ok variant in some cases so upgrading to a quiche with
that change would cause some issues. This change turns that specific
error into the Ok(0) case which overnet can handle. This works with the
current version of quiche and will allow a future upgrade to a version
which includes that change.

Change-Id: I1ccc9393b366ccfdf41573fa80b71e373ae59b0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580652
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>