[netemul] Shutdown realm before continuing with test teardown
This test was prone to flakes because sometimes the dhcpv6-client would
still be in the middle of starting up when the test finished. Test
interfaces get removed when the fuchsia.netemul.network/Endpoint channel
to them is dropped, and this sometimes meant an interface was removed in
between dhcpv6-client creating a socket and trying to bind it to that
interface. This race led to unexpected error logs that caused test
failures.

In order to prevent realm shutdown from racing with test interface
teardown, block on realm shutdown before the channel handles go out of
scope.

Fixed: 82913
Test: fx test netemul-v2-tests
Multiply: netstack-management-integration-test
Change-Id: If5672bba263610099f4e8c542a8893082c22bcac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576922
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>