[audio] Fix conversion warnings
We forgot to file our own bug for this. The original bug was
fxbug.dev/58162, which is now closed.

These warnings revealed a bug in OutputProducer, which was not correctly
truncating SIGNED_24_IN_32 values. This bug has been fixed. Fixing this
bug requires adding an operation (either a shift-by-constant or
logical-and-by-constant). Some profiling with audio_mixer_profiler shows
that the impact of this change is basically negligible.

Change-Id: I6adab9787731c1e3126f5abb81e2d5a19b8f2ba7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542986
Fuchsia-Auto-Submit: Tom Bergan <tombergan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>