[fx] Remove use of ffx from list-packages
list-packages reports information about what packages were built by
default, not what packages are in the repo. In trying to "move to ffx",
we changed the semantics by having it list packages in the repository.
This reverts that change pending a UX review.

Fixed: 83465

Change-Id: Ibc9e491929abed1ed1629710af14a024a295b060
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573564
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>