[camera][ge2d] Fix wrong exported format index
The wrong format index is exported at the output of the GE2D
stage of Sherlock configuration 0. The setting of the format
value by SetResolution will only set the input when the mode
is the overlay (in-place) mode. This means that the output
resolution is actually the input resolution, but since only
the input mode is set, the output format index is left as an
invalid value.

Please note that this invalid GE2D output format value is
never used in any Sherlock configurations so there is
technically no impact.

Bug: 197900636
Test: Run normal unit tests to make sure there is no
      accidental side effect.
Change-Id: I3fb1b8b467fb0cd4ca9c41dc7534e8663d39a715
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578187
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ernest Hua <ernesthua@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>