[msd-arm-mali] Cancel perf count reads when in protected mode
HWCPipe triggers a performance counter read when starting. If the GPU
happened to be in protected mode at that point (e.g. due to an earlier
test running) the trigger wouldn't occur and the read would hang
forever.

To fix this, immediately cancel existing triggers and mark them as
completed when entering protected mode, and do the same to any new
triggers when in protected mode.

Test: msd_arm_mali_nonhardware_tests
Test: mali_vulkan_test
Fixed: 83455
Fixed: 84449

Change-Id: I87fe0425243ffe5d33991377a4b473f26b60a09c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581321
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: John Bauman <jbauman@google.com>