[io] Require fuchsia.io to have explicit oridinals
Previously, the FIDL compiler had some magic to rename fuchsia.io
ordinals to fuchsia.io1. This magic was intended to make it easier to
migrate to fuchsia.io2 by letting the new io2 versions of the methods
have the "normal" ordinal.

Now that we are in the midst of the transition, we don't want to be
confused by this magic. This CL requires that all ordinals in fuchsia.io
be written explicitly in the protocol definition, removing the magic. To
ensure that we do not miss anything, the compiler now crashes whenever
it would have previously applied the magic.

Bug: 77623
Change-Id: I1380a114879119ac8b309fc787cd706b82675de7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581401
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: James Robinson <jamesr@google.com>