[fdio][zxio] Remove the is_socket check, add tests
The fdio and zxio libraries had an internal check of whether an object
instance was of a socket type or a different type. This dates back to
the original introduction of dispatch tables for I/O operations and is
currently only used to compute error codes for a few operations Fuchsia
does not support. These operations are not tested, as shown by
https://fuchsia-review.googlesource.com/c/fuchsia/+/565561.

This adds tests for these error codes and removes the "is a socket"
check. In practice, this means we now generate ENOSYS for sendmmsg /
recvmmsg instead of ENOTSOCK, and the zxio interface is a bit simpler.

Bug: 43267
Change-Id: I8dedb06f5b6d74fdba83e87d3bddd66810c23e71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580859
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>