[wayland] Flatland scale and crop.
Refactor image import to support multiple surfaces
and flatland instances. This also enables crop and
scale support as that's trivial when we can create
unique image instances for each surface.

Note: A bug in Sommelier is preventing cropping from
working as expected. The source rectangle is not
correctly adjusted for the buffer scale. weston-scaler
can be used to verify that this works as expected when
that bug has been fixed.

Bug: 78911
Change-Id: I8f6380ac29c2d86d496532754f036e13ef74810d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581562
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>