[clock][testrunner] Fix race condition in testrunner test.
There is a race condition between when After() is called and when
Advance() is called on the fakeClock. Advance() should be called after
After() is called so that it advances the timer that After() adds to the
fakeClock.

Fixed: 84466
Change-Id: I841c1e10c4167d8b7190c065d891115727d3bc1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580929
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>