[tools] Update the host Linux sysroot to Debian 11 "buster"
We follow Chromium with our host Linux sysroot version. Chromium has
switched to sysroot Debian 11 "buster" in M93 which was released on
Aug 31, 2021 so we should be able to update sysroot for Fuchsia as well.

Note that this change doesn't yet update the sysroot itself. Once this
change lands, the lock file will be picked up the linux-sysroot builder
which will build the new sysroot. This can then be rolled into prebuilts.

Bug: 64200
Change-Id: I4ddc19a548f5b1f70100364998aaf49f857be411
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581723
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Petr Hosek <phosek@google.com>