[netsvc] Temporarily disable netsvc stress tests on ASAN builder
FIDL is undergoing a wire format migration. As part of this, a
transformer is used to transform between the old and new wire
format before/after decode/encode. This slows down the netsvc
stress test so that it keeps timing out.

This CL disables the test on ASAN builds until the migration is
complete.

Bug: 84697
Change-Id: I18069a58ddb240fd734f292cb0a04ba3c6b1168e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581902
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>