[scrutiny][plugins] Static packages plugin
This change introduces a new plugin that collects the static packages
list from a build. It depends on the devmgr_config plugin to load the
build's pkgsvr parameters, which contains the system image hash. The
system image hash is the starting point for hash-following to the
system's static packages list.

Bug: 81536
Test: fx test scrutiny
Test: verify_static_pkgs rule on allowlisted product builds

Change-Id: Ic3e2e2bf48e9f6746129153456468a54dc1583ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578987
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>