[kernel] Send MP_IPI_INTERRUPT when Killing or Suspending threads
When killing or suspending a thread, issuing an MP_IPI_RESCHEDULE is
unnecessary.  We don't need to cause the target CPU to call into the
scheduler.  We only need it to take an interrupt because the interrupt
handler will check and process thread signals as needed.

Change-Id: I998dbcf32839af447ff5f60e9d205fd4ac6b94dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581381
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>