[llcpp] Define message encoding / decoding classes outside of type class
The class definitions of structs, tables, requests and responses are
made busier than necessary because they contain the definitions of
UnownedEncodedMessage, OwnedEncodedMessage, DecodedMessage within them.
This makes it harder to find and read the core parts of the type
definition.

This CL moves these definitions outside of the body in order to make the
classes easier to read.

Fixed: 84425
Change-Id: Iac694eb846cbb1778aaf8fc5ca0952a3bfd9ac5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582262
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>