[f2fs] Handles corruption on a file name
This CL returns NOT_FOUND when an inode block has a wrong lenth
of its filename. In this case, it sets kBad flag and clears link for
the successful deletion of the vnode without actual purging on disk.
It will inform of the corruption when fsck supports a repair feature.

Fixed: 84681

Change-Id: I35086c45f25fc623bbaa5944f13e4140e1803720
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582623
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>