[net-types] Prefer Witness::get over into_addr
For `Witness<A>` types, when both `get` and `into_addr` are available
(which is the case when `A: Copy`), prefer `get`. Document this
preference in the docs for `Witness` and update call sites to respect
this preference.

Also remove "into" methods on other types where possible.

Bug: 83331

Change-Id: Ic3928b5fa336ccc02ac65ede8db36a0770d5015b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573843
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>