[f2fs] Add range checkers for checkpoint packs
This CL adds checking the validity for some fields that
a checkpoint pack have such as the sizes of nat/sit version bitmaps
and the next free node id. If they are invalid, it stops mounting
since it means that there is corruption on the read-only superblock
or the checkpoint pack.

Fixed: 84609, 84611

Change-Id: I2807a8d413128c3ba2ee938b4d89c5c41d26fa67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582621
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>