[fidlc] Sometimes we have no ordinals on a method
Due to previous compilation errors, we might not have an ordinal on a
method, and must therefore protect the access.

Implementing the restriction on `fuchsia.io` requiring explicit ordinals
as a compiler check, rather than an assert. This avoids the compiler
failing in a more obscure way, and is more in line with how fidlc is
meant to operate.

Bug: 77623
Change-Id: Ic69ed0edfffcd99d30921fa56abb4c5ec2cfc4ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582341
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Pascal-Louis Perez <pascallouis@google.com>