[fidl][llcpp] Async methods unbind on send errors
When a client method call encounters an error, it should lead to binding
teardown. We already do this for errors in the receiving path; this CL
extends that to cover errors in the sending path as well.

This CL started out simple but uncovered two subtle issues:

1. If the teardown is immediately initiated on any error, that could
lead to racy discarding of unread messages. The chosen behavior is to
drain all messages when the error is peer closed. The full details are
explained in a comment block in `llcpp_client_base.cc`. This would be a
good case to formalize and validate in a future dynamic binding spec
conformance test.

2. The LLCPP client supports both async and sync calls. We would like to
make the sync calls initiate teardown on errors too, but that turned out
much more tricky because the sync calls use the C++17 guaranteed return
value optimization to avoid any copies or moves of the result type. As
a result, the code performing the sync call cannot check the result of
the call because it is a pr-value built directly on the user's stack.
Furthermore, if a sync call drags out long after the client object is
destroyed, it may cause use-after-free as it tries to teardown a
non-existent client.

It's likely still possible to make sync calls in the async client
initiate teardown but that could involve more changes. This CL sets the
foundation via some refactoring.

Fixed: 78989

Change-Id: I215bcc46dd4979eee192cbc54cfac21f660f6069
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582013
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>