[appmgr] Temporary fix for cache purging
Appmgr's notion of free space is not correctly computed for FVM-enabled
filesystems, and doesn't take into account the budgeted maximum size of
the partition, so cache purging might never kick in (since the
denominator is higher than expected, and we purge when using 95% or more
of the partition).

This CL is a temporary fix for this issue which addresses the problem
with minimal churn.  A proper fix is impending but requires plumbing
much more information through the system, so this fix will take up the
slack in the meantime.

This fix is only enabled if the new GN argument
`minfs_board_maximum_data_bytes` is set.  This CL does not set this GN
argument for any devices.

Bug: 83978
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582190
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Change-Id: I2bdf9096485769197b66c9f460a8101a4d180fe8