[test_manager] Allow test resources to extend realm lifetime
Today, the lifetime of a test realm is loosely tied to
actions taken by a test executor. Specifically, the Archivist under a
test realm must remain alive long enough for the executor to drain logs
collected from the test realm.
We'll soon have custom artifacts, which will be exported via a handle
to storage accessed by a component in the test realm. Test manager will
need to keep realms alive while an executor is accessing the storage,
otherwise, component manager is free to delete the storage. Unlike
logs, the handles for accessing custom artifacts must be created after
the test terminates.

This change provides the `run_test' method with a mutable reference to
the running suite, which allows it to store signals that can be used to
keep the realms alive. It also moves the creation of log tasks into
'run_test'.

Bug: 75690
Testing: Pure refactor, test with existing integ tests
Multiply: test_manager_test

Change-Id: I04af51d85968bb13d1da3dae2407c74a87dcefc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580306
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>