[cts] Update verify_release script
Now written in python and moved to
//sdk/cts/build/scripts/verify_release.py.

This can be run as a test or standalone as a script, though this likely
will never be able to execute as-is in CQ since it runs fx commands.

Following CLs will add a step to run the tests at the end, and reset
your "fx args" settings to be less disruptive.

Bug: 84049
Change-Id: I6d885394d0aa63eaf1d5487d647be87cf2350507
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573941
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>