[virtualization][vsock] Avoid sending spurious shutdowns to endpoints
virtio-vsock responds to unexpected incoming packets by replying with
a RESET response. To implement this, the receive path allocates a new
NullConnection and places it in the SHUTDOWN state. When the transmit
path of virtio-vsock processes this NullConnection, it will send out
a RESET packet and destroy the NullConnection.

However, when the transmit path destroys this connection, it also
notifies any connected endpoint that a socket has been destroyed. The
guest manager reacts to this event by freeing the port associated with
the connection.

In the case of an unexpected packet being received, however, no port was
actually ever allocated: this leads to the guest manager freeing a port
that may be in use by an unrelated socket.

Update virtio-vsock to avoid notifying endpoints and closed
NullConnections.

Bug: fxbug.dev/84286

Change-Id: Ifac0bd75dca50b4b0c32bcc79b0ef7e69310f51c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582625
Commit-Queue: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>