[starnix] Prevent "rm foo/.." and "rmdir foo/.."
Previously, we would look up the child twice in the unlink codepath,
once in NamespaceNode and again in the DirEntry code. We need to perform
the lookup exactly once in case another thread is modifying the file
system concurrently. This CL centralizes the lookup in the DirEntry
code.

In order to handle mount points correctly, we need to plumb the
NamespaceNode for the parent to the DirEntry code. This CL also follows
the same pattern for rename.

This code is still slightly incorrect because we do not hold the lock on
the mount point table for the duration of these functions. Another
thread could potentially make the directory entry into a mount point
while we are trying to unlink or rename the entry. This CL includes a
TOOD for that issue, which will need to be resolved in a future CL.

Bug: 79440
Test: CannotRemoveDots
Change-Id: Ibdab59613d7d35e180d18b95b4fc01fe09e6cfae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582851
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>