[graphics]: Remove zxtest from fake device tree.
zxtest assertion were used with the only effect of providing formatted
output into stdout.

This usage pattern is in conflict with strict zxtest, which requires a
running test context to propagate assertions and fail the running test.

Usage was replaced with equivalent log statements with return in the
case of assert.

Bug: 83770
Test: CQ
Change-Id: Iaeec925c9e712146e23e1e4b4d0f5b6a9af14945
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582041
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>