[kernel][spinlock] Implement strong CMPX semantics for trylock
The current compare-and-exchange operation used in the ARM64 trylock
implementation has "weak" semantics instead of "strong" semantics.
This means that there are two distinct ways that the operation can
fail:

1) It can fail because the spinlock state is observed to be contested
   by another CPU.
2) It can fail because the store exclusive operation fails due to
   either the local or global monitor not being in the exclusive state.

If the operation had "strong" semantics, then only #1 could cause the
function to return; a STXR failure would cause the operation to try
again.  While the rules for the global monitor dropping to the clear
state require implementations to guarantee a transition from exclusive
-> clear any time another PE attempts a STXR to the same physical
address, they allow implementations to transition either the global or
local monitor drop out (or not) for a very long list of other reasons.
Effectively, the store operation can spontaneously fail at almost any
time.

Typically, this distinction does not matter.  In any situation where
the lock _may_ be contended, the operation _could_ fail, and so the
difference between reasons #1 and #2 really should not matter.  There
are a small number of cases, however, where it might.

In this case, consider the test code in bug 83983.  The test is
testing a path where there is a try acquire operation, and the test
code needs to guarantee that this operation succeeds as part of its
functionality testing.  It _should_ be able to provide this guarantee,
since it can guarantee that there are no other CPUs contending for the
lock in question during the test.  Currently, however, with weak
semantics, this behavior is not a guarantee.  Instead, the operation
can fail at any time for any reason.

One option would be to inject a behavior change into the code being
tested, causing it to perform an unconditional acquire of the lock
during testing, while using the trylock implementation in production.
Another is to provide strong semantics to the trylock operation
itself.  This CL implements the second option.

Finally, this CL also adds a CLREX to the end of the trylock
implementation.  Previously, in the case that the trylock operation
fails because it actually does observe contention, the global/local
monitors were being left in the exclusive state.  Now, that state is
cleared explicitly.

I can currently think of only one way in which this difference might
end up causing a change in behavior, and I cannot think of a
particular reason why it should matter.  Specifically, if a CPU
performs an LDXR on physical address A, and then aborts the trylock
because it observes contention, and the monitor is left in a exclusive
state, then a second STXR to local A which occurs _after_ the trylock
operation _could_ succeed when it _should_ have failed.

This "spurious" success would almost certainly imply a bug somewhere
else in the system.  Exclusive stores should never be issued without
an associated exclusive load, so to have a random STXR take place
after a failed trylock operation should pretty much always be an
error.  That said, adding the clear is simple, should not harm
anything, and aligns this hand-written assembly implementation more
with the CMPX implementation generated by clang.

Fixed: 83983
Change-Id: I940957db25521258776c55ef0f5f31cb19bb4c53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578321
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: John Grossman <johngro@google.com>