[lockdep] Reorder lockdep bookkeeping operations
Change the order of operation when an instrumented lock is released
while lockdep is enabled. Instead of releasing the lock and then
updating the bookkeeping, first update the bookkeeping and then
release the lock.

For details, see http://fxb/84827.  Also add a regression test which
repro'ed the issue before, and no longer does.

Fixed: 84827
Change-Id: I1989278abb777a46b8212f201cf9cad2ea1ffdc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582469
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>