[fidlc] Add type_shape_v* to type field of JSON IR
There isn't currently a universal way to get type sizing information for
types in the JSON IR. Bindings may redo the logic by combining size
information on specific decl types with special logic handling arrays,
vectors, etc.

In particular, this is useful for implementing different logic on
members of unions and tables based on the type size. Currently, it is
necessary to reimplement some of the type shape logic to get this
information. When type shapese are in the IR under each type, it is much
more straightforward to extract it.

Bug: 79578
Change-Id: Ib5ddf64954c3ed5ad232b79d7da5ff3e8552da21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582663
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>