[fx] Delete legacy_set feature
It has been 3 months since we switched `fx set` to use the Go
implementation by default (fxrev.dev/529261) so it's time to delete the
old implementation as promised.

Also replace the Bash unit tests for the old version of `fx set` with
tests for the logic that determines whether to rebuild the fx-set Go
tool. That is the only interesting Bash logic that remains; all other
behavior is covered by the Go tool's unit tests.

Bug: 68465
Change-Id: If1cdc87a2ff89e93621bc3e8ecbce3c878830976
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582223
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>