[fs_test][minfs] Disable resize test in ASAN builds
A previous CL
I8ee4577bd0d5172934401d599ae5bea0774796c4 attempted to disable only the
minfs resize test on ASAN builds. The motivation for this was to prevent
timeouts due to a temporary change in FIDL for a wire format migration
that slows down communication.

While the previous CL appeared to work due to flakiness, it didn't
actually fix the problem and there are still flaky failures. This CL
disables ASAN testing for a broader number of resize tests, but actually
works (no failures from the test on the ASAN builder 10x in a row).

Bug: 84697
Change-Id: I858fa51dd44941c23b09d4d904a60c7bf1600376
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/583461
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>