[kernel][ktrace] Refactor lib/ktrace for testing.
In particular, turn the ktrace_state struct into a proper C++ class
with methods which do not assume that they operate on the global
singleton instance.  Rewire the existing API (things like
ktrace_enabled or ktrace_probe) to be inline passthrus to the global
singleton.

This, along with a just a couple of code-injection modifications (not
present in this CL) should allow instantiation a ktrace system
different from the global singleton, which should allow us to write
some independent unit tests.

Change-Id: Id886fb6fb057409859dc627807fa5ffa59c2b91e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579743
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>